requirements.txt
streamlit
chromadb
sentence-transformers
requests
pyyaml
pypdf
